requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0
